added icon button in story book, and done some changes.
