+++ src/App.js
@@ -1,13 +1,13 @@
 import "./App.scss";
 import { ListContainer } from "./components/ListContainer/listContainer";
-import { Navbar } from "./components/Navbar/Navbar";
+import { PrjHeader } from "./components/PrjHeaderComponent/PrjHeader";
 
 
 
 function App() {
   return (
     <div className="App">
-      <Navbar />
+      <PrjHeader />
       <ListContainer />
     </div>
   );
