+++ src/components/ContentComponents/Content.scss
@@ -13,6 +13,7 @@
     margin: 1% 0% 0.5% 0%;
 }
 
+
 .add-power-ups-button {
     background-color: transparent;
     color: #172b4d;
