+++ src/components/HeaderLabel/Label.scss
@@ -19,7 +19,7 @@
 }
 
 .label-list{
-    width: 100%;
+    width: 300px;
     height: 50vh;
     overflow-y: scroll;
 }
@@ -49,6 +49,7 @@
     height: 20px;
     border-radius: 50%;
     margin: 7px;
-    // display: flex;
     align-self: center;
+    box-shadow: none;
+    border-color: transparent;
 }
