+++ src/components/IconButtonComponent/IconButton.scss
@@ -1,0 +1,3 @@
+.button-icon-container {
+    margin-right: 10%;
+}
