+++ src/components/PrjHeaderComponent/PrjHeader.js
@@ -1,6 +1,6 @@
 import { Button } from "antd";
 import React, { useState } from "react";
-import "./Navbar.scss";
+import "./PrjHeader.scss";
 import { Typography } from "antd";
 import { StarOutlined, StarFilled } from "@ant-design/icons";
 import { IoMdContacts } from "react-icons/io";
@@ -9,17 +9,21 @@ import { TfiAngleDown } from "react-icons/tfi";
 import { SlCalender } from "react-icons/sl";
 import { IoIosJet } from "react-icons/io";
 import { ImPower } from "react-icons/im";
-import { FilterComponent } from "../NavbarFilter/Filter";
 import { FaAngleDoubleUp } from "react-icons/fa";
-import {BsThreeDots,BsPersonPlus} from 'react-icons/bs'
+import { BsThreeDots, BsPersonPlus } from "react-icons/bs";
+import { HeaderFilter } from "../HeaderFilter/Filter";
 
-export const Navbar = () => {
+export const PrjHeader = () => {
   const { Text } = Typography;
 
   const [isFillStar, SetIsFillStar] = useState(false);
 
-  const favouriteButton = () => {
+  const addFavourite = () => {
     SetIsFillStar(true);
+  };
+
+  const removeFavourite = () => {
+    SetIsFillStar(false);
   };
 
   return (
@@ -28,21 +32,23 @@ export const Navbar = () => {
         <section className="header-container">
           <div className="title-container">
             <Text className="title">Project Management</Text>
-            <Button onClick={favouriteButton} className="button">
-              {isFillStar ? (
-                <StarFilled className="star-icon" />
-              ) : (
-                <StarOutlined className="star-icon1" />
-              )}
-            </Button>
+            {isFillStar ? (
+              <Button onClick={removeFavourite} className="button">
+                <StarFilled className="star-yellow" />
+              </Button>
+            ) : (
+              <Button onClick={addFavourite} className="button">
+                <StarOutlined className="star-white" />
+              </Button>
+            )}
           </div>
-          <div className="line-container"> | </div>
+          <div className="line-container"></div>
           <div className="workspace-container">
             <Button className="workspace-button">
               <IoMdContacts className="workspace-icon" /> Workspace visible
             </Button>
           </div>
-          <div className="line-container"> | </div>
+          <div className="line-container"> </div>
           <div className="board-container">
             <Button className="board-button">
               <MdLeaderboard className="board-icon" /> Board
@@ -60,11 +66,11 @@ export const Navbar = () => {
               <ImPower className="calender-icon" /> Automation
             </Button>
           </div>
-          <div className="line-container"> | </div>
+          <div className="line-container">  </div>
           <div className="filter-container">
-            <FilterComponent />
+            <HeaderFilter />
           </div>
-          <div className="line-container"> | </div>
+          <div className="line-container">  </div>
           <div className="avtar">
             <FaAngleDoubleUp className="double-arrow-icon" />
             DB
@@ -72,11 +78,13 @@ export const Navbar = () => {
           <div className="avtar1">SP</div>
           <div>
             <Button className="share-button">
-            <BsPersonPlus className="person-icon"/> Share
+              <BsPersonPlus className="person-icon" /> Share
             </Button>
           </div>
-          <div className="line-container"> | </div>
-          <Button className="more-option-button"> <BsThreeDots/> </Button>
+          <div className="line-container">  </div>
+          <Button className="more-option-button">
+            <BsThreeDots />
+          </Button>
         </section>
       </nav>
     </>
